Use size_t

+++ lib/str.c
@@ -2110,7 +2110,7 @@ grn_text_lltoa(grn_ctx *ctx, grn_obj *buf, long long int i)
     char *curr = GRN_BULK_CURR(buf);
     char *tail = GRN_BULK_TAIL(buf);
     if (grn_lltoa(i, curr, tail, &curr)) {
-      unsigned int new_size;
+      size_t new_size;
       new_size = grn_bulk_compute_new_size(ctx,
                                            buf,
                                            GRN_BULK_WSIZE(buf) + UNIT_SIZE);
